requirements.txt
tensorflow-cpu==2.20.0
numpy
pillow